requirements.txt
torch>=2.2.0
torchvision>=0.15.2
transformers>=4.30.2
pillow>=9.5.0
streamlit>=1.24.0
pydicom>=2.3.1
opencv-python>=4.7.0.72
numpy>=1.24.3
pandas>=2.0.2
scikit-learn>=1.2.2
matplotlib>=3.7.1
monai>=1.2.0
langchain>=0.0.235
sentence-transformers>=2.2.2
chromadb>=0.4.6
python-dotenv>=1.0.0 